Removed extended warnings (#2)
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@17361 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: 4318675a0cd5173ac786f42f93b15c162d39fdbe

+++ spatial/rcp.viewer.stream.map/src/de/uniol/inf/is/odysseus/rcp/viewer/stream/map/LayerUpdater.java
@@ -169,7 +169,9 @@ public class LayerUpdater extends ArrayList<ILayer> implements
 			// This deletes the oldest element. If more than one element
 			// is in this time-window, more than one element will
 			// be deleted.
-			puffer.purgeElementsBefore(deleteTime.plus(1));
+			if( deleteTime != null ) {
+				puffer.purgeElementsBefore(deleteTime.plus(1));
+			}
 
 			// Update "current-list", timeSlider and all the other things which
 			// rely on the startTimeStamp
